fixes fixes gccixes (builds on 4.x again)
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@2155 406a0c4d-033a-0410-8de8-e80135713968

+++ src/tests.cpp
@@ -1220,7 +1220,7 @@ void BenchSort ()
 	DWORD * pKeys = new DWORD [ MAXSIZE ];
 	SortPayload_t * pValues = new SortPayload_t [ MAXSIZE ];
 
-	for ( int iGen=0; iGen<sizeof(g_dSortDataGens)/sizeof(g_dSortDataGens[0]); iGen++ )
+	for ( int iGen=0; iGen<(int)(sizeof(g_dSortDataGens)/sizeof(g_dSortDataGens[0])); iGen++ )
 		for ( int iLen=MINSIZE; iLen<=MAXSIZE; iLen*=10 )
 	{
 		int64_t tmSort = 0;
